requirements.txt
mysql-connector-python>=8.0.33
tabulate>=0.9.0


